put signals handling in its own module.
