+++ circus/workers.py
@@ -12,17 +12,6 @@ from circus.controller import Controller
 
 class Workers(object):
 
-    WORKERS = {}
-
-    SIGNALS = map(
-        lambda x: getattr(signal, "SIG%s" % x),
-        "HUP QUIT INT TERM WINCH".split()
-    )
-
-    SIG_NAMES = dict(
-        (getattr(signal, name), name[3:].lower()) for name in dir(signal)
-    )
-
     def __init__(self, num_workers, cmd, check_delay, warmup_delay, endpoint):
         self.cmd = cmd
         self.num_workers = num_workers
@@ -30,50 +19,23 @@ class Workers(object):
         self.warmup_delay = warmup_delay
         self.ctrl = Controller(endpoint, self, self.check_delay)
         self.pid = os.getpid()
-
-        # set zmq socket
-        self.ctx = zmq.Context()
-        self.skt = self.ctx.socket(zmq.REQ)
-        self.skt.connect(endpoint)
         self.worker_age = 0
-
-        # init signals
-        map(lambda s: signal.signal(s, self.signal), self.SIGNALS)
-        signal.signal(signal.SIGCHLD, self.handle_chld)
+        self.WORKERS = {}
 
         print "Starting master on pid %s" % self.pid
 
-    def signal(self, sig, frame):
-        if sig in self.SIG_NAMES:
-            signame = self.SIG_NAMES.get(sig)
-            self.skt.send(signame)
-
-
-    def handle_chld(self, *args):
-        pass
-
-    def handle_hup(self):
+    def handle_reload(self):
         pass
 
     def handle_quit(self):
         self.halt()
 
-    def handle_int(self):
-        self.halt()
-
     def run(self):
         self.manage_workers()
         while True:
-            try:
-                self.reap_workers()
-                self.manage_workers()
-
-                self.ctrl.poll()
-            except KeyboardInterrupt:
-                self.halt()
-            except SystemExit:
-                raise
-
+            self.reap_workers()
+            self.manage_workers()
+            self.ctrl.poll()
 
     def reap_workers(self):
         for wid, worker in self.WORKERS.items():
@@ -124,10 +86,5 @@ class Workers(object):
 
     def terminate(self):
         self.kill_workers()
-
-        try:
-            self.ctx.destroy(0)
-            self.ctrl.terminate()
-        except:
-            pass
+        self.ctrl.terminate()
